:sparkles: Start to move out the settings and default parsing and use reflection
