+++ pkg/cli/cmds.go
@@ -1,0 +1,1 @@
+package cli
