+++ pkg/cmds/cobra_test.go
@@ -14,7 +14,7 @@ func TestAddZeroArguments(t *testing.T) {
 	desc := CommandDescription{
 		Arguments: []*Parameter{},
 	}
-	err := AddArguments(cmd, &desc)
+	err := AddArguments(cmd, desc.Arguments)
 	// assert that err is nil
 	require.Nil(t, err)
 }
@@ -30,7 +30,7 @@ func TestAddSingleRequiredArgument(t *testing.T) {
 			},
 		},
 	}
-	err := AddArguments(cmd, &desc)
+	err := AddArguments(cmd, desc.Arguments)
 	require.Nil(t, err)
 	assert.Nil(t, cmd.Args(cmd, []string{"bar"}))
 	assert.Error(t, cmd.Args(cmd, []string{}))
@@ -64,7 +64,7 @@ func TestAddTwoRequiredArguments(t *testing.T) {
 			},
 		},
 	}
-	err := AddArguments(cmd, &desc)
+	err := AddArguments(cmd, desc.Arguments)
 	require.Nil(t, err)
 	assert.Nil(t, cmd.Args(cmd, []string{"bar", "foo"}))
 	assert.Error(t, cmd.Args(cmd, []string{}))
@@ -103,7 +103,7 @@ func TestOneRequiredOneOptionalArgument(t *testing.T) {
 			},
 		},
 	}
-	err := AddArguments(cmd, &desc)
+	err := AddArguments(cmd, desc.Arguments)
 	require.Nil(t, err)
 	assert.Nil(t, cmd.Args(cmd, []string{"bar", "foo"}))
 	assert.Nil(t, cmd.Args(cmd, []string{"foo"}))
@@ -140,7 +140,7 @@ func TestOneOptionalArgument(t *testing.T) {
 			},
 		},
 	}
-	err := AddArguments(cmd, &desc)
+	err := AddArguments(cmd, desc.Arguments)
 	require.Nil(t, err)
 	assert.Error(t, cmd.Args(cmd, []string{"bar", "foo"}))
 	assert.Nil(t, cmd.Args(cmd, []string{"foo"}))
@@ -168,7 +168,7 @@ func TestDefaultIntValue(t *testing.T) {
 			},
 		},
 	}
-	err := AddArguments(cmd, &desc)
+	err := AddArguments(cmd, desc.Arguments)
 	require.Nil(t, err)
 	values, err := GatherArguments([]string{}, desc.Arguments, false)
 	require.Nil(t, err)
@@ -302,7 +302,7 @@ func TestInvalidDefaultValue(t *testing.T) {
 				},
 			},
 		}
-		err := AddArguments(cmd, &desc)
+		err := AddArguments(cmd, desc.Arguments)
 		if err == nil {
 			t.Errorf("Expected error for type %s and value %v\n", failingType.Type, failingType.Value)
 		}
@@ -322,7 +322,7 @@ func TestTwoOptionalArguments(t *testing.T) {
 			},
 		},
 	}
-	err := AddArguments(cmd, &desc)
+	err := AddArguments(cmd, desc.Arguments)
 	require.Nil(t, err)
 	assert.Error(t, cmd.Args(cmd, []string{"bar", "foo", "blop"}))
 	assert.Nil(t, cmd.Args(cmd, []string{"bar", "foo"}))
@@ -343,7 +343,7 @@ func TestFailAddingRequiredAfterOptional(t *testing.T) {
 			},
 		},
 	}
-	err := AddArguments(cmd, &desc)
+	err := AddArguments(cmd, desc.Arguments)
 	assert.Error(t, err)
 }
 
@@ -358,7 +358,7 @@ func TestAddStringListRequiredArgument(t *testing.T) {
 			},
 		},
 	}
-	err := AddArguments(cmd, &desc)
+	err := AddArguments(cmd, desc.Arguments)
 	require.Nil(t, err)
 	assert.Nil(t, cmd.Args(cmd, []string{"bar", "foo"}))
 	assert.Error(t, cmd.Args(cmd, []string{}))
@@ -377,7 +377,7 @@ func TestAddStringListOptionalArgument(t *testing.T) {
 			},
 		},
 	}
-	err := AddArguments(cmd, &desc)
+	err := AddArguments(cmd, desc.Arguments)
 	require.Nil(t, err)
 	assert.Nil(t, cmd.Args(cmd, []string{"bar", "foo"}))
 	assert.Nil(t, cmd.Args(cmd, []string{"foo"}))
@@ -409,7 +409,7 @@ func TestFailAddingArgumentAfterStringList(t *testing.T) {
 			},
 		},
 	}
-	err := AddArguments(cmd, &desc)
+	err := AddArguments(cmd, desc.Arguments)
 	assert.Error(t, err)
 }
 
@@ -424,7 +424,7 @@ func TestAddIntegerListRequiredArgument(t *testing.T) {
 			},
 		},
 	}
-	err := AddArguments(cmd, &desc)
+	err := AddArguments(cmd, desc.Arguments)
 	require.Nil(t, err)
 	assert.Nil(t, cmd.Args(cmd, []string{"1", "2"}))
 	assert.Error(t, cmd.Args(cmd, []string{}))
@@ -447,7 +447,7 @@ func TestAddStringListRequiredAfterRequiredArgument(t *testing.T) {
 			},
 		},
 	}
-	err := AddArguments(cmd, &desc)
+	err := AddArguments(cmd, desc.Arguments)
 	require.Nil(t, err)
 	assert.Nil(t, cmd.Args(cmd, []string{"foo", "bar"}))
 	assert.Error(t, cmd.Args(cmd, []string{}))
@@ -470,7 +470,7 @@ func TestAddStringListOptionalAfterRequiredArgument(t *testing.T) {
 			},
 		},
 	}
-	err := AddArguments(cmd, &desc)
+	err := AddArguments(cmd, desc.Arguments)
 	require.Nil(t, err)
 	assert.Nil(t, cmd.Args(cmd, []string{"foo", "bar", "baz"}))
 	assert.Nil(t, cmd.Args(cmd, []string{"foo", "bar"}))
@@ -494,7 +494,7 @@ func TestAddStringListOptionalAfterOptionalArgument(t *testing.T) {
 			},
 		},
 	}
-	err := AddArguments(cmd, &desc)
+	err := AddArguments(cmd, desc.Arguments)
 	require.Nil(t, err)
 	assert.Nil(t, cmd.Args(cmd, []string{"foo", "bar", "baz"}))
 	assert.Nil(t, cmd.Args(cmd, []string{"foo", "bar"}))
@@ -516,7 +516,7 @@ func TestAddStringListRequiredAfterOptionalArgument(t *testing.T) {
 			},
 		},
 	}
-	err := AddArguments(cmd, &desc)
+	err := AddArguments(cmd, desc.Arguments)
 	assert.Error(t, err)
 }
 
@@ -603,9 +603,9 @@ func testCommandParseHelper(t *testing.T, desc CommandDescription, expected *exp
 		},
 	}
 
-	err := AddArguments(cmd, &desc)
+	err := AddArguments(cmd, desc.Arguments)
 	require.Nil(t, err)
-	err = AddFlags(cmd, &desc)
+	err = AddFlags(cmd, desc.Flags)
 	require.Nil(t, err)
 	cmd.SetArgs(expected.Args)
